+++ bin/lint.sh
@@ -5,4 +5,4 @@ set -e
 bin_folder="`dirname "$0"`"
 root_folder="$bin_folder/.."
 
-"$bin_folder/HOST_PATH/flux-js-lint" "$root_folder"
+"$bin_folder/HOST_PATH/js-lint" "$root_folder"